player_reid/requirements.txt
opencv-python
numpy
torch
PyYAML